+++ CHANGES.rst
@@ -10,6 +10,7 @@ Unreleased changes
 
 *   Do not scan the filesystem for files to compress
     when all compression algorithms are disabled.
+*   Guarantee that file are compressed after all other plugins have run.